uml: sparsify: Skip test since it needs qcow2.

+++ sparsify/test-virt-sparsify.sh
@@ -19,6 +19,11 @@
 export LANG=C
 set -e
 
+if [ "$(../fish/guestfish get-backend)" = "uml" ]; then
+    echo "$0: skipping test because uml backend does not support qcow2"
+    exit 77
+fi
+
 rm -f test-virt-sparsify-1.img test-virt-sparsify-2.img
 
 # Create a filesystem, fill it with data, then delete the data.  Then
